Added Service interface and implementation

+++ src/main/java/com/akshat/fooddelivery/model/FoodItems.java
@@ -4,7 +4,6 @@ import com.sun.istack.NotNull;
 import lombok.*;
 
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.validation.constraints.NotBlank;
 
@@ -16,7 +15,7 @@ import javax.validation.constraints.NotBlank;
 @Builder
 public class FoodItems {
     @Id
-    private Long itemId;
+    private int itemId;
     @NotNull
     private Long restaurantId;
     @NotBlank
